# server/public/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Hola API
  version: "1.0.0"
servers:
  - url: http://localhost:{port}
    description: Local server
    variables:
      port:
        default: "3000"
paths:
  /api/apps:
    post:
      summary: Deploy an application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                appName:
                  type: string
                  description: Name of the application to deploy
                version:
                  type: string
                  description: Version of the application to deploy (default is "latest")
              required:
                - appName
      responses:
        "200":
          description: App deployment initiated (SSE stream)
          content:
            text/event-stream:
              schema:
                type: string
        "400":
          description: Missing app name
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
    get:
      summary: List all deployed applications
      responses:
        "200":
          description: List of applications retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  apps:
                    type: array
                    items:
                      type: string
        "500":
          description: Failed to list applications
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  details:
                    type: string
  /api/apps/deploy:
    post:
      summary: Alternative endpoint to deploy an application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                appName:
                  type: string
                  description: Name of the application to deploy
                version:
                  type: string
                  description: Version of the application to deploy (default is "latest")
              required:
                - appName
      responses:
        "200":
          description: App deployment initiated (SSE stream)
          content:
            text/event-stream:
              schema:
                type: string
        "400":
          description: Missing app name
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /api/apps/{appName}/details:
    get:
      summary: Get details about a deployed application
      parameters:
        - name: appName
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Application details retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  appName:
                    type: string
                  status:
                    type: string
                  config:
                    type: object
                  files:
                    type: array
                    items:
                      type: string
        "404":
          description: Application not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        "500":
          description: Failed to get application details
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  details:
                    type: string
  /api/apps/{appName}/upgrade:
    post:
      summary: Upgrade an application to a new version
      parameters:
        - name: appName
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                version:
                  type: string
                  description: Version to upgrade to (default is "latest")
      responses:
        "200":
          description: Upgrade initiated (SSE stream)
          content:
            text/event-stream:
              schema:
                type: string
        "500":
          description: Failed to upgrade application
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /api/apps/{appName}:
    delete:
      summary: Remove a deployed application
      parameters:
        - name: appName
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Application removal initiated (SSE stream)
          content:
            text/event-stream:
              schema:
                type: string
        "500":
          description: Failed to remove application
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /api/apps/{appName}/files:
    post:
      summary: Upload a file for an application
      parameters:
        - name: appName
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                filePath:
                  type: string
              required:
                - file
                - filePath
      responses:
        "201":
          description: File uploaded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  path:
                    type: string
        "400":
          description: Missing fields or file / invalid filePath
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        "500":
          description: File upload failed
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    
    get:
      summary: List all files for an application
      parameters:
        - name: appName
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Files retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  files:
                    type: array
                    items:
                      type: object
                      properties:
                        path:
                          type: string
                        name:
                          type: string
                  count:
                    type: integer
        "400":
          description: Invalid app name
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        "500":
          description: Failed to list files
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
...
  /api/apps/{appName}/start:
    post:
      summary: Start an application
      parameters:
        - name: appName
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Application start initiated (SSE stream)
          content:
            text/event-stream:
              schema:
                type: string
        "500":
          description: Failed to start application
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /api/apps/{appName}/stop:
    post:
      summary: Stop an application
      parameters:
        - name: appName
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Application stop initiated (SSE stream)
          content:
            text/event-stream:
              schema:
                type: string
        "500":
          description: Failed to stop application
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /api/apps/{appName}/restart:
    post:
      summary: Restart an application
      parameters:
        - name: appName
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Application restart initiated (SSE stream)
          content:
            text/event-stream:
              schema:
                type: string
        "500":
          description: Failed to restart application
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /api/apps/{appName}/backup:
    post:
      summary: Create a backup of an application
      parameters:
        - name: appName
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Backup created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  backupId:
                    type: string
                  message:
                    type: string
        "404":
          description: Application not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        "500":
          description: Failed to create backup
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /api/apps/{appName}/backups:
    get:
      summary: List all backups for an application
      parameters:
        - name: appName
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Backups retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  backups:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        createdAt:
                          type: string
                        size:
                          type: number
                  count:
                    type: integer
        "404":
          description: Application not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        "500":
          description: Failed to list backups
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /api/apps/{appName}/backup/{backupId}:
    get:
      summary: Get details about a specific backup
      parameters:
        - name: appName
          in: path
          required: true
          schema:
            type: string
        - name: backupId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Backup details retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  backupId:
                    type: string
                  createdAt:
                    type: string
                  size:
                    type: number
                  metadata:
                    type: object
        "404":
          description: Backup not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        "500":
          description: Failed to get backup details
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /api/apps/{appName}/restore/{backupId}:
    post:
      summary: Restore an application from a backup
      parameters:
        - name: appName
          in: path
          required: true
          schema:
            type: string
        - name: backupId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Application restoration initiated (SSE stream)
          content:
            text/event-stream:
              schema:
                type: string
        "404":
          description: Backup not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        "500":
          description: Failed to restore application
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /api/apps/{appName}/logs:
    get:
      summary: Get logs for an application
      parameters:
        - name: appName
          in: path
          required: true
          schema:
            type: string
        - name: lines
          in: query
          required: false
          schema:
            type: integer
            default: 100
        - name: service
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Logs retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  logs:
                    type: array
                    items:
                      type: string
        "404":
          description: Application not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        "500":
          description: Failed to get logs
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
...
  /api/apps/{appName}/events:
    get:
      summary: Stream events for an application (SSE)
      parameters:
        - name: appName
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: SSE event stream established
          content:
            text/event-stream:
              schema:
                type: string
        "404":
          description: Application not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string

  # Configuration Management Endpoints
  /api/config:
    get:
      summary: Retrieve system-wide configuration
      parameters:
        - name: key
          in: query
          required: false
          schema:
            type: string
          description: Specific configuration key to retrieve
      responses:
        "200":
          description: Configuration retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  config:
                    type: object
        "500":
          description: Failed to retrieve configuration
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  details:
                    type: string

    post:
      summary: Create or update multiple system configuration values
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                config:
                  type: object
              required:
                - config
      responses:
        "200":
          description: Configuration updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  config:
                    type: object
                  message:
                    type: string
        "500":
          description: Failed to update configuration
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  details:
                    type: string

    delete:
      summary: Delete multiple system configuration values
      parameters:
        - name: keys
          in: query
          required: true
          schema:
            type: string
          description: Comma-separated list of keys to delete
      responses:
        "200":
          description: Configuration values deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  deletedKeys:
                    type: array
                    items:
                      type: string
                  message:
                    type: string
        "400":
          description: Missing keys parameter
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        "500":
          description: Failed to delete configuration values
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  details:
                    type: string
  /api/config/{key}:
    put:
      summary: Create or update a specific system configuration value
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                value:
                  type: object
              required:
                - value
      responses:
        "200":
          description: Configuration value updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  key:
                    type: string
                  value:
                    type: object
                  message:
                    type: string
        "500":
          description: Failed to update configuration value
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  details:
                    type: string
                    
    delete:
      summary: Delete a specific system configuration value
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Configuration value deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  key:
                    type: string
                  message:
                    type: string
        "404":
          description: Configuration key not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        "500":
          description: Failed to delete configuration value
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  details:
                    type: string
  /api/config/{appName}:
    get:
      summary: Retrieve application-specific configuration
      parameters:
        - name: appName
          in: path
          required: true
          schema:
            type: string
        - name: key
          in: query
          required: false
          schema:
            type: string
          description: Specific configuration key to retrieve
      responses:
        "200":
          description: Application configuration retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  appName:
                    type: string
                  config:
                    type: object
        "404":
          description: Application not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        "500":
          description: Failed to retrieve configuration
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  details:
                    type: string
                    
    post:
      summary: Create or update multiple app configuration values
      parameters:
        - name: appName
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                config:
                  type: object
              required:
                - config
      responses:
        "200":
          description: Application configuration updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  appName:
                    type: string
                  config:
                    type: object
                  message:
                    type: string
        "404":
          description: Application not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        "500":
          description: Failed to update configuration
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  details:
                    type: string
                    
    delete:
      summary: Delete application configuration
      parameters:
        - name: appName
          in: path
          required: true
          schema:
            type: string
        - name: keys
          in: query
          required: false
          schema:
            type: string
          description: Comma-separated list of keys to delete (if omitted, all keys will be deleted)
      responses:
        "200":
          description: Configuration deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  appName:
                    type: string
                  message:
                    type: string
                  deletedKeys:
                    type: array
                    items:
                      type: string
        "404":
          description: Application not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        "500":
          description: Failed to delete configuration
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  details:
                    type: string
...