# server/src/routes.ts
const express = require("express");
import { Application } from "express";
const appsController = require("./controllers/apps");
const filesController = require("./controllers/files");
const configController = require("./controllers/config"); // This controller needs to be created

/**
 * Registers all API routes on the given Express application
 *
 * @param app - The Express application instance
 */
export const registerRoutes = (app: Application): void => {
  // Health check endpoint
  app.get("/", (req, res) => {
    res.json({
      status: "ok",
      version: process.env.npm_package_version || "1.0.0",
    });
  });

  // Application management routes
  app.post("/api/apps/deploy", appsController.deployApp); // Deploy a new application
  app.get("/api/apps", appsController.listApps); // List all deployed applications
  app.get("/api/apps/:appName/details", appsController.getAppDetails); // Get details about an application
  app.post("/api/apps/:appName/upgrade", appsController.upgradeApp); // Upgrade an existing application
  app.delete("/api/apps/:appName", appsController.removeApp); // Remove an application
  app.post("/api/apps/:appName/start", appsController.startApp); // Start an application
  app.post("/api/apps/:appName/stop", appsController.stopApp); // Stop an application

  // Missing app management routes (to be implemented)
  app.post("/api/apps/:appName/restart", appsController.restartApp); // Restart an application

  // Backup & Restore routes (to be implemented)
  app.post("/api/apps/:appName/backup", appsController.createBackup); // Create a backup
  app.get("/api/apps/:appName/backups", appsController.listBackups); // List all backups
  app.get(
    "/api/apps/:appName/backup/:backupId",
    appsController.getBackupDetails
  ); // Get backup details
  app.post(
    "/api/apps/:appName/restore/:backupId",
    appsController.restoreFromBackup
  ); // Restore from backup

  // Logs & Monitoring routes (to be implemented)
  app.get("/api/apps/:appName/logs", appsController.getAppLogs); // Get application logs
  app.get("/api/apps/:appName/metrics", appsController.getAppMetrics); // Get application metrics
  app.get("/api/apps/:appName/health", appsController.getAppHealth); // Check application health

  // File management routes
  app.post(
    "/api/apps/:appName/files",
    filesController.uploadFile,
    filesController.handleFileUpload
  ); // Upload a file
  app.get("/api/apps/:appName/files", filesController.listFiles); // List files for an application
  app.get("/api/apps/:appName/files/:filePath(*)", filesController.getFile); // Get a specific file
  app.delete(
    "/api/apps/:appName/files/:filePath(*)",
    filesController.deleteFile
  ); // Delete a specific file

  // Configuration management routes
  // System configuration routes
  app.get("/api/config", configController.getSystemConfig); // Get all system configuration
  app.post("/api/config", configController.createSystemConfig); // Create/update multiple system config values
  app.put("/api/config/:key", configController.updateSystemConfigValue); // Create/update a specific system config value
  app.delete("/api/config/:key", configController.deleteSystemConfigValue); // Delete a specific system config value
  // Add this route handler that checks for the keys query parameter
  app.delete("/api/config", (req, res) => {
    if (req.query.keys) {
      return configController.deleteMultipleSystemConfigValues(req, res);
    }
    return res.status(400).json({ error: "Missing keys query parameter" });
  });

  // App configuration routes
  app.get("/api/config/:appName", configController.getAppConfig); // Get all app configuration
  app.post("/api/config/:appName", configController.createAppConfig); // Create/update multiple app config values
  app.put("/api/config/:appName/:key", configController.updateAppConfigValue); // Create/update a specific app config value
  app.delete(
    "/api/config/:appName/:key",
    configController.deleteAppConfigValue
  ); // Delete a specific app config value
  // Update this route handler to check for the keys query parameter
  app.delete("/api/config/:appName", (req, res) => {
    if (req.query.keys) {
      return configController.deleteMultipleAppConfigValues(req, res);
    }
    return configController.deleteAppConfig(req, res);
  });

  // Encrypted app configuration routes
  app.get(
    "/api/config/:appName/encrypted",
    configController.getAppEncryptedConfig
  ); // Get all encrypted app configuration
  app.post(
    "/api/config/:appName/encrypted",
    configController.createAppEncryptedConfig
  ); // Create/update multiple encrypted values
  app.put(
    "/api/config/:appName/encrypted/:key",
    configController.updateAppEncryptedValue
  ); // Create/update a specific encrypted value
  app.delete(
    "/api/config/:appName/encrypted/:key",
    configController.deleteAppEncryptedValue
  ); // Delete a specific encrypted value

  // Events stream route (SSE for real-time updates)
  app.get("/api/apps/:appName/events", appsController.streamEvents); // Stream events for an app
};

// Export using CommonJS syntax
module.exports = {
  registerRoutes,
};